requirements.txt
tkinter
tkhtmlview
markdown
openai
faiss-cpu
numpy
tiktoken
python-dotenv
Pillow